+++ tetris/include/tetrisRunner.h
@@ -3,7 +3,7 @@
 
 #include "tetris.h"
 
-void initTetrisGame(int (*boardPtr)[BOARDSIZE]);
+void initTetrisGame(uint8_t (*boardPtr)[BOARDSIZE]);
 enum gameSignal handleTetrisTick();
 enum gameSignal handleTetrisUserEvent(char c);
 bool gameOverTetrisAnimation();
